Fix subject cleanup in confluent base IT

+++ integration-tests/src/test/java/io/apicurio/tests/ConfluentBaseIT.java
@@ -68,17 +68,29 @@ public abstract class ConfluentBaseIT extends ApicurioRegistryBaseIT {
     }
 
     protected void clearAllConfluentSubjects() throws IOException, RestClientException {
-        for (String confluentSubject : confluentService.getAllSubjects()) {
-            logger.info("Deleting confluent schema {}", confluentSubject);
-            try {
-                confluentService.deleteSubject(confluentSubject);
-            } catch (RestClientException e) {
-                if (e.getStatus() == 404) {
-                    //subjects may be already deleted
-                    continue;
-                }
-                throw e;
+        int breakerCount = 0;
+        boolean broken = false;
+        while (confluentService != null && confluentService.getAllSubjects() != null) {
+            if (breakerCount > 5) {
+                broken = true;
+                break;
             }
+            for (String confluentSubject : confluentService.getAllSubjects()) {
+                logger.info("Deleting confluent schema {}", confluentSubject);
+                try {
+                    confluentService.deleteSubject(confluentSubject);
+                } catch (RestClientException e) {
+                    if (e.getStatus() == 404 || e.getStatus() == 422) {
+                        // subject may be already deleted or have references
+                        continue;
+                    }
+                    throw e;
+                }
+            }
+            breakerCount++;
+        }
+        if (broken) {
+            throw new IOException("Failed to delete ALL confluent schemas.");
         }
     }
